+++ CHANGELOG.md
@@ -20,7 +20,7 @@ This is a log of major user-visible changes in each phpMyFAQ release.
 - added possibility to enable/disable admin session counter (Thorsten)
 - added EU Data Act 2025 compliance (Thorsten)
 - added Urdu translation (Thorsten)
-- added dark mode / light mode toggle (Thorsten)
+- added dark mode / light mode / high contrast toggle (Thorsten)
 - added support for PHP 8.5 (Thorsten)
 - added support for .env files (Thorsten)
 - added support for OpenSearch (Thorsten)